refactor(bspline): move _find_span and _count_multiplicity to _bspline_knots
These are general-purpose knot query primitives that belong alongside
the other knot utilities rather than in the slice-specific module.
